Modified Delete.php for session login

+++ basic_project_backbone-master/public_html/api/drinks/delete.php
@@ -4,17 +4,26 @@ require '../../../bootloader.php';
 
 $drinksModel = new App\Drinks\Model();
 $drinks = $drinksModel->get(['id' => intval($_POST['id'])]);
-if (!$drinks) {
+if($_SESSION){
+   if (!$drinks) {
     print json_encode([
         'status' => 'fail',
         'errors' => [
             'Drink doesn`t exist'
         ]
     ]);
+    } else {
+        $drinksModel->delete($drinks[0]);
+        print json_encode([
+            'status' => 'success',
+            'errors' => []
+        ]);
+    } 
 } else {
-    $drinksModel->delete($drinks[0]);
     print json_encode([
         'status' => 'success',
-        'errors' => []
+        'errors' => [
+            'You need to login'
+        ]
     ]);
 }
